server/requirements.txt
flask
flask-cors
rasterio
numpy
matplotlib
seaborn
scikit-learn
xgboost
tensorflow
pandas
Pillow
scikit-image
